+++ Changelog.md
@@ -13,6 +13,8 @@ Fixes:
 - For attribute lists, make sure to collapse the whitespace.
 - When attributes have an `@XmlSerialName` annotation with a default namespace
   value, then this will result in a non-qualified attribute.
+- Fix nullable QName serialization (and probably other nullable inline-like)
+  serialization.
 
 # 0.86.1
 *(July 5, 2023)<br />*